+++ src/api.js
@@ -26,14 +26,15 @@ export default class API {
   updateTask(id, data) {
     const headers = new Headers();
     headers.append(`Authorization`, this._authorization);
+    headers.append(`Content-Type`, `application/json`);
 
     return fetch(`https://11.ecmascript.pages.academy/task-manager/tasks/${id}`, {
       method: `PUT`,
-      body: JSON.stringify(data),
+      body: JSON.stringify(data.toRAW()),
       headers,
     })
       .then(checkStatus)
       .then((response) => response.json())
-      .then(taskModel.parseTasks);
+      .then(taskModel.parseTask);
   }
 }
