+++ src/components/task-edit-component.js
@@ -149,9 +149,8 @@ export default class TaskEditComponent extends AbstractSmartComponent {
 
   getData() {
     const form = this.getElement().querySelector(`.card__form`);
-    const formData = new FormData(form);
 
-    return this._parseFormData(formData);
+    return new FormData(form);
   }
 
   _applyFlatpickr() {
@@ -242,24 +241,6 @@ export default class TaskEditComponent extends AbstractSmartComponent {
         >`
       );
     }).join(`\n`);
-  }
-
-  _parseFormData(formData) {
-    const repeatingDays = DAYS.reduce((acc, day) => {
-      acc[day] = false;
-      return acc;
-    }, {});
-    const date = formData.get(`date`);
-
-    return {
-      description: formData.get(`text`),
-      color: formData.get(`color`),
-      dueDate: date ? new Date(date) : null,
-      repeatingDays: formData.getAll(`repeat`).reduce((acc, it) => {
-        acc[it] = true;
-        return acc;
-      }, repeatingDays),
-    };
   }
 
   _isAllowableDescriptionLength(description) {
